requirements_test.txt
pytest-homeassistant-custom-component
asyncssh==2.22.0
aiofiles==25.1.0
